backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy[asyncio]>=2.0.25
pydantic>=2.5.3
passlib>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.0
email-validator>=2.0.0
mailjet-rest>=1.3.4
python-dotenv>=1.0.0
h5py

# Celery & Pillow
celery[redis]>=5.3.0,<6.0.0
redis
aioredis
pillow>=10.3.0,<11.0.0

# Asynchronous database drivers
aiosqlite
aiomysql

# Password hashing
argon2-cffi

# JOSE: dùng pycryptodome backend để tránh pyasn1
python-jose[pycryptodome]>=3.3.0
pyasn1==0.4.8
pyasn1-modules>=0.2.8